requirements.txt
# Core Frameworks and Tools
Celery
Django
django-extensions
ipython

# Third-Party Libraries and Packages
django-admin-rangefilter
django-debug-toolbar
django-crispy-forms
django-filter
django-import-export
django-phonenumber-field
beautifulsoup4
crispy_bootstrap5
phonenumbers
requests

# REST APIs and Serialization
djangorestframework
djangorestframework-simplejwt
drf-yasg

# REST API renderers
djangorestframework-xml
djangorestframework-yaml

# Storage and File Handling
boto3
django-storages
Pillow

# Code Quality and Linting
flake8
flake8-builtins
flake8-print

# Testing
pytest
pytest-django
pytest-cov
pytest-mock
model_bakery
